+++ src/replication.c
@@ -526,7 +526,7 @@ void syncWithMaster(aeEventLoop *el, int fd, void *privdata, int mask) {
     while(maxtries--) {
 #ifdef _WIN32
         snprintf(tmpfile,256,
-            "temp-%d.%lld.rdb",(int)time(NULL),(long long int)getpid());
+            "temp-%lld.%lld.rdb",(long long)time(NULL),(long long)getpid());
         dfd = open(tmpfile,O_CREAT|O_WRONLY|O_EXCL|O_BINARY,_S_IREAD|_S_IWRITE);
 #else
         snprintf(tmpfile,256,
